requirements.txt
streamlit==1.34.0
pandas==2.2.2
fpdf2==2.7.7
openpyxl==3.1.2
Pillow==10.3.0